signature: merge types from async-signature
